Make fixes a bit more robust

+++ .scripts/setup_fixes_php.sh
@@ -5,11 +5,15 @@ IFS=$'\n\t'
 setup_fixes_php()
 {
     info "PHP fixes"
-    info "- Installing new php7.3-fpm"
-    export DEBIAN_FRONTEND=noninteractive
-    export UCF_FORCE_CONFFNEW=1
-    apt-get -y -o Dpkg::Options::=--force-confnew install php7.3-fpm
-    export DEBIAN_FRONTEND=
-    export UCF_FORCE_CONFFNEW=
+    if [[ $(dpkg -l | grep -c php7.3-fpm) == 0 ]]; then
+        info "- Installing new php7.3-fpm"
+        export DEBIAN_FRONTEND=noninteractive
+        export UCF_FORCE_CONFFNEW=1
+        apt-get -y -o Dpkg::Options::=--force-confnew install php7.3-fpm
+        export DEBIAN_FRONTEND=
+        export UCF_FORCE_CONFFNEW=
+    else
+        info "- php7.3-fpm already fixed!"
+    fi
     info "- Done"
 }
